implement simulated mouse buttons for Mac laptops similar to plan9port.
For systems without a three button mouse, the keyboard modifier keys
can be used to modify the effect of the main mouse button. On Mac
systems, the option key changes the main button to button 2, and the
Command key changes it to button 3.

Pressing the key after the button is held down adds the button to
form a chord, so that for example on Macs selecting text with the
trackpad button and then typing Option without letting go of the button
will cause a 1-2 chord, cutting the selection.

These changes were inspired by rsc's plan9port.